v2.0.0: Changed __type__ value syntax
In order to support nested class definitions, the `__type__` value now has the following
syntax:
`my.module.path::My.Class.Path`
So module specification is separated from class specification by `::` instead of
the common `.`.
This is necessary in order to make the split of the fully-qualified name at the
correct point.